pyproject.toml
[project]
name = "gtm-advisor"
version = "0.1.0"
description = "Self-Service Agentic AI Platform for GTM Advisory - Singapore SMEs"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "GTM Advisor Team" }]
keywords = ["ai", "agents", "gtm", "marketing", "lead-generation", "singapore"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "httpx>=0.27",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
    "alembic>=1.18.2",
    "aiosqlite>=0.22.1",
    "slowapi>=0.1.9",
    "apscheduler>=3.11.2",
    "pypdf>=6.8.0",
    "python-docx>=1.2.0",
    "intelligence",
    "vector-store",
    "knowledge",
    "openai>=2.15.0",
]

[tool.uv.sources]
intelligence = { workspace = true }
vector-store = { workspace = true }
knowledge = { workspace = true }

[project.optional-dependencies]
mcp = [
    "beautifulsoup4>=4.12",
    "feedparser>=6.0",
    "lxml>=5.0",
]
llm = [
    "openai>=1.30",
    "anthropic>=0.25",
    "instructor>=1.0",
]
perplexity = [
    "openai>=1.30",  # Perplexity uses OpenAI-compatible API
]
newsapi = [
    "newsapi-python>=0.2",
]
eodhd = [
    "eodhd>=1.0",
]
crm = [
    "hubspot-api-client>=8.0",
    "simple-salesforce>=1.12",
    "msal>=1.24",
]
email = [
    "sendgrid>=6.10",
]
api = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.0",
]
redis = [
    "redis>=5.0",
]
postgres = [
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
]
observability = [
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-jaeger>=1.21",
    "prometheus-client>=0.20",
]
vectors = [
    "qdrant-client>=1.9",
    "sentence-transformers>=2.7",
]
all = [
    "gtm-advisor[mcp,llm,perplexity,newsapi,eodhd,crm,email,api,redis,postgres,observability,vectors]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "respx>=0.21",  # HTTP mocking
]

[project.scripts]
gtm-advisor = "services.gateway.src.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages", "agents", "services"]

[tool.uv.workspace]
members = ["packages/*", "agents/*", "services/*"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["packages", "agents", "services", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Function call in default argument (FastAPI Depends)
    "B904",   # Raise from inside except
    "E712",   # Comparison to True/False (required for SQLAlchemy column expressions)
    "UP046",  # Generic class uses Generic subclass (stylistic preference)
    "SIM102", # Use single if statement (not always clearer)
    "SIM103", # Return condition directly (not always clearer)
    "SIM105", # Use contextlib.suppress (debatable)
    "SIM108", # Use ternary operator (not always clearer)
    "SIM110", # Use any() instead of loop (not always clearer)
    "C416",   # Unnecessary comprehension (sometimes clearer)
]

[tool.ruff.lint.per-file-ignores]
# Agent classes must match base class method signatures
"agents/*/src/*.py" = ["ARG002"]
# Test fixtures may be used only for side effects
"tests/**/*.py" = ["ARG002"]
# Tool classes may have stub methods with unused args
"packages/tools/src/*.py" = ["ARG002"]
# Scripts need path setup before imports
"scripts/*.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["packages", "agents", "services"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow-running tests",
    "network: Tests requiring network access",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["newsapi.*", "eodhd.*", "hubspot.*", "simple_salesforce.*", "msal.*", "sendgrid.*"]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
